pyproject.toml
[project]
name = "eqx-bot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.124.4",
    "openai>=2.12.0",
    "opentelemetry-api>=1.39.1",
    "opentelemetry-instrumentation-openai>=0.50.1",
    "opentelemetry-sdk>=1.39.1",
    "python-dotenv>=1.2.1",
    "pyyaml>=6.0.3",
    "streamlit>=1.52.2",
    "uvicorn>=0.38.0",
    "weaviate-client>=4.18.3",
]
